requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
asn1crypto==1.5.1
attrs==25.4.0
bcrypt==4.3.0
bleach==6.3.0
boolean.py==5.0
boto3==1.42.39
botocore==1.42.54
build==1.4.0
CacheControl==0.14.4
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.4
cryptography==46.0.5
cyclonedx-python-lib==11.6.0
Deprecated==1.3.1
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
filelock==3.24.3
frozenlist==1.8.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-cloud-core==2.5.0
google-cloud-vision==3.12.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gotrue==2.12.4
grpcio==1.78.1
grpcio-status==1.78.1
h11==0.16.0
h2==4.3.0
hatchling==1.28.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
jmespath==1.1.0
license-expression==30.4.4
limits==5.8.0
Mako==1.3.10
mcp==1.26.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
oauthlib==3.3.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.7
packageurl-python==0.17.6
packaging==25.0
passlib==1.7.4
pathspec==1.0.4
pillow==12.1.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.2
pluggy==1.6.0
postgrest==2.28.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
py-serializable==2.1.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.3.2
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-asyncio==0.26.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-magic==0.4.27 ; sys_platform != 'win32'
python-magic-bin==0.4.14 ; sys_platform == 'win32'
python-multipart==0.0.22
PyYAML==6.0.3
realtime==2.28.0
redis==7.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
ruff==0.15.2
s3transfer==0.16.0
scramp==1.4.8
sentry-sdk==2.53.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.50.0
storage3==2.28.0
stripe==14.3.0
supabase==2.28.0
supafunc==0.10.2
tomli==2.4.0
tomli_w==1.2.0
trove-classifiers==2026.1.14.14
types-bleach==6.3.0.20251115
types-cachetools==6.2.0.20251022
types-cffi==1.17.0.20250915
types-html5lib==1.1.11.20251117
types-Pillow==10.2.0.20240822
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-requests==2.32.4.20260107
types-setuptools==82.0.0.20260210
types-webencodings==0.5.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
webencodings==0.5.1
websockets==14.2
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
